+++ lib/resources/asset_paths.dart
@@ -3,5 +3,7 @@ const _graphics = "$basePath/graphics";
 const _data = "$basePath/data";
 
 const largeLogo = "$_graphics/3.0x/logo.png";
+const smallLogo = "$_graphics/logo.png";
 const colorNames = "$_data/color_names.json";
 const colorSuggestions = "$_data/color_suggestions.json";
+const palettes = "$_data/palettes.json";