+++ lib/sources/app/styling/blur/blur_scheme.dart
@@ -1,0 +1,5 @@
+import 'package:colored/sources/app/styling/blur/blur_sigma.dart';
+
+abstract class BlurScheme {
+  BlurSigma get medium;
+}