+++ lib/sources/domain/view_models/names_list/names_list_injector.dart
@@ -1,8 +1,9 @@
 import 'dart:async';
 
+import 'package:colored/sources/data/services/map_filter/color_names_filter.dart';
 import 'package:colored/sources/data/services/memoizer/default_memoizer.dart';
 import 'package:colored/sources/data/services/names/color_names_service.dart';
-import 'package:colored/sources/data/services/data_loader/color_names_data_loader.dart';
+import 'package:colored/sources/data/services/data_loader/color_names_loader.dart';
 import 'package:colored/sources/domain/view_models/names_list/names_list_state.dart';
 import 'package:colored/sources/domain/view_models/names_list/names_list_view_model.dart';
 
@@ -15,9 +16,10 @@ class NamesListInjector {
       NamesListViewModel(
         stateController: stateController ?? StreamController<NamesListState>(),
         namesService: ColorNamesService(
-          dataLoader: ColorNamesDataLoader(
+          dataLoader: ColorNamesLoader(
             memoizer: DefaultMemoizer<Map<String, String>>(),
           ),
+          filter: const ColorNamesFilter(),
         ),
       );
 }