+++ lib/sources/presentation/layouts/naming/naming_layout.dart
@@ -3,23 +3,19 @@ import 'package:colored/sources/presentation/layouts/naming/naming_cross_fade_te
 import 'package:colored/sources/presentation/layouts/picker/picker_tab_bar.dart';
 import 'package:flutter/material.dart';
 
-class NamingLayout extends StatelessWidget {
-  const NamingLayout({this.onSearchPressed, Key key}) : super(key: key);
-
-  final void Function() onSearchPressed;
+class NamingLayout extends StatelessWidget implements PreferredSizeWidget {
+  const NamingLayout({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final localization = Localization.of(context).converter;
     return AppBar(
       title: NamingCrossFadeText(defaultText: localization.colorConverter),
-      actions: <Widget>[
-        IconButton(
-          icon: const Icon(Icons.search),
-          onPressed: onSearchPressed,
-        ),
-      ],
       bottom: const PickerTabBar(),
     );
   }
+
+  @override
+  Size get preferredSize =>
+      const Size.fromHeight(kToolbarHeight + kTextTabBarHeight);
 }