+++ lib/sources/presentation/widgets/containers/overlay_container.dart
@@ -1,7 +1,8 @@
+import 'dart:ui';
+
+import 'package:colored/sources/app/styling/blur/blur_data.dart';
 import 'package:colored/sources/app/styling/elevation/elevation_data.dart';
 import 'package:colored/sources/app/styling/opacity/opacity_data.dart';
-import 'package:colored/sources/app/styling/padding/padding_data.dart';
-import 'package:colored/sources/app/styling/padding/padding_scheme.dart';
 import 'package:colored/sources/app/styling/radii/radius_data.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -9,52 +10,38 @@ import 'package:flutter/material.dart';
 class OverlayContainer extends StatelessWidget {
   const OverlayContainer({
     this.child,
-    this.padding,
+    this.elevation,
+    this.borderRadius,
     Key key,
   }) : super(key: key);
 
   final Widget child;
-  final EdgeInsets padding;
+  final double elevation;
+  final BorderRadius borderRadius;
 
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
     final colors = theme.colorScheme;
-    final elevation = ElevationData.of(context).elevationScheme;
+    final elevationScheme = ElevationData.of(context).elevationScheme;
     final radii = RadiusData.of(context).radiiScheme;
-    final padding = PaddingData.of(context).paddingScheme;
-    final deviceSafeArea = MediaQuery.of(context).padding;
     final opacity = OpacityData.of(context).opacityScheme;
-    return SafeArea(
-      child: OrientationBuilder(
-        builder: (_, orientation) => Padding(
-          padding: _getOuterPadding(orientation, deviceSafeArea, padding),
-          child: Material(
-            elevation: elevation.low,
+    final blur = BlurData.of(context).blurScheme.medium;
+    final borderRadius = this.borderRadius ?? BorderRadius.all(radii.large);
+    return Material(
+      elevation: elevation ?? elevationScheme.low,
+      color: Colors.transparent,
+      borderRadius: borderRadius,
+      child: ClipRRect(
+        borderRadius: borderRadius,
+        child: BackdropFilter(
+          filter: ImageFilter.blur(sigmaX: blur.x, sigmaY: blur.y),
+          child: Container(
             color: colors.primary.withOpacity(opacity.overlay),
-            borderRadius: BorderRadius.all(radii.large),
             child: child,
           ),
         ),
       ),
     );
-  }
-
-  EdgeInsets _getOuterPadding(
-    Orientation orientation,
-    EdgeInsets deviceSafeArea,
-    PaddingScheme paddingScheme,
-  ) {
-    if (kIsWeb) {
-      return paddingScheme.large;
-    } else {
-      if (orientation == Orientation.portrait) {
-        return deviceSafeArea.bottom == 0
-            ? paddingScheme.small
-            : EdgeInsets.symmetric(horizontal: deviceSafeArea.bottom / 2);
-      } else {
-        return paddingScheme.small;
-      }
-    }
   }
 }